refactor: Move download button to global components
